Beer_server/requirements.txt
asgiref==3.6.0
attrs==22.1.0
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
